refactor: some stuff
